# docs/swagger/url-shortener.yml
openapi: 3.0.1
info:
  title: URL Shortener
  description: URL Shortening API - Coded in Kotlin - Powered by Micronaut
paths:
  /uba/tdl/metrics/{hash}:
    get:
      tags:
      - Internal URL Operations
      summary: Retrieves URL usage metrics.
      operationId: getMetrics
      parameters:
      - name: hash
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: getMetrics 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/URLMetrics'
  /uba/tdl/redirect/{hash}:
    post:
      tags:
        - Public URL Operations
      summary: Redirects shortened URL to it's original domain.
      operationId: redirectURL
      parameters:
        - name: hash
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: redirectURL 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /uba/tdl/search:
    post:
      tags:
        - Internal URL Operations
      summary: Retrieves any shortened site that matches certain keywords.
      operationId: searchSites
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Keywords'
        required: true
      responses:
        "200":
          description: searchSites 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
  /uba/tdl/shorten:
    post:
      tags:
        - Public URL Operations
      summary: Shortens a given URL.
      operationId: shortenURL
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LongURL'
        required: true
      responses:
        "200":
          description: shortenURL 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortURL'
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          nullable: true
    Keywords:
      required:
        - values
      type: object
      properties:
        values:
          minItems: 1
          type: array
          items:
            type: string
    LongURL:
      required:
        - url
      type: object
      properties:
        url:
          minLength: 1
          type: string
    SearchResult:
      required:
        - matchingWords
        - url
      type: object
      properties:
        url:
          minLength: 1
          type: string
        matchingWords:
          type: array
          items:
            type: string
    ShortURL:
      required:
        - url
      type: object
      properties:
        url:
          minLength: 1
          type: string
    URLMetrics:
      required:
        - lastUsage
        - originalURL
        - shortURL
        - usageCount
      type: object
      properties:
        originalURL:
          minLength: 1
          type: string
        shortURL:
          minLength: 1
          type: string
        usageCount:
          minimum: 0
          type: integer
          format: int32
        lastUsage:
          type: string
          format: date-time